+++ CHANGELOG.md
@@ -7,3 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Independent recovery analysis [#5](https://github.com/IN-CORE/pyincore-incubator/tree/5-lisa-new-independent-recovery-analysis)
+